Abstract logic for signing
